Fixed configurations (up to Stanza). Alphabetized the methods in Context.
Configurations is now a separate collection, and ConfCollection is now ConfigurationFile. Stanza has the new
constructor signature, but is otherwise untouched (i.e., still needs to be fixed). Added IllegalOperation for operations
which are implied by the superclass but known not to function.

Refactored @context to @service everywhere in the test suite in preparation for making Context into Service and
removing the distinction between them.