Added bulletproofing to test suite.
Made Context#restart always set a restart_required message so users can check for it.
Added rescue Errno::ECONNRESET to several places involving restarts and checks for
server state (since that is a legitimate response while the server is going down for restarts).
Added is_ready? checks to job creation in test_jobs. Made sure that jobs using sleep commands
are deleted and that we wait long enough for the command to actually terminate so Splunk can
cleanly delete the sleep_command app.