Documented and refactored export results parsing.
Renamed classes to be MultiResultsReader and PuppetResultsReader.
Added docstrings and reordered file.
Added spool method to ResultsReader which skips the rest of this results set.
Added spool after each yield in MultiResultsReader so subsequent results readers
will be in a sane state even if a previous one was not completely read.