Finished resolving group review comments.
Added namespace argument to create on Configurations and ConfigurationFile.
Removed gratuitous setting of path in Index#attach.
Added instance field for sid on Job and changed Job#sid to use it. This lets sid be available before the Job is ready.
Renamed the fields returned by Job#timeline and documented them.
SavedSearch#history reuses the states it is returned to create its Jobs rather than fetching them from the server.
Job#update checks if search is a required field before setting it (to avoid race condition on newer versions of Splunk where it is unnecessary).
Changed create_stream to create_export in test_jobs.