Add Pry based shell
Create SplunkShell class, child of Service.
Create methods to execute a oneshot,run a shell, give basic help.
More advanced than the simple CLI as we can operate with returns
of different types and extend the shell as discussed in #26 on the
main GitHub repo - https://github.com/splunk/splunk-sdk-ruby.